requirements.txt
GitPython==3.1.0
matplotlib==3.1.1
numpy==1.17.3
opencv-python==4.2.0.32
Pillow==6.2.1
pytz==2019.3
scikit-image==0.17.2
scipy==1.4.1
tensorboard==2.0.0
torch==1.3.0
torchvision==0.4.1
tqdm==4.43.0
